detect/http_stat_code: move tests into tests/
